Move cloneChainsConfig into the test that uses it
Addresses review comment https://github.com/nomiclabs/hardhat/pull/1928#discussion_r751599069
